Updates to the EWW bar.
Multiple updates to the bar. Started working on the scss portion of the
bar. Styling is difficult, and I've never been good at css. Ugh. This is
probably gonna take a while. Anyway, here's what got done:

- Added styling for the workspace buttons.
- Attempted styling for the volume scale.
- Added temporary button for killing the bar widget for ease of
  development.

The colorscheme was taken from one of the examples at elkowar/eww. I
can't be bothered to check which one. rixhern or something.
	modified:   bar/eww.scss
	modified:   bar/eww.yuck