Weird decisions from like two years ago
I'm a bit confused by my decisions to split up the hotkeys for selecting
workspaces. It was a strange decision to make in retrospect, but I
suppose it makes a little sense thinking back. If I'm remembering
correctly, I figured it would be easier to reach the 1-4 keys with my
left and 7-0 keys with my right hands. It was unintuitive at best, so I
just changed it.