Removed close-bar widget, added Power widget
The close-bar widget served its purpose in testing, but it's been
replaced by a hotkey in my sxhkd config, so I can remove it without much
concern. I only really needed it when I was using Gnome, anyway.

I added a Power widget so I wouldn't have to pop open a terminal for
poweroff, reboot and hibernate. I've included suspend in there too, for
all I don't really need it. The problem is, it doesn't actually work.
Anytime I try to add it to bar-content and refresh eww the bar simply
refuses to render. I'll need to look into it, but the only thing I can
think to do is check the inspector. If that doesn't work I'll have to
submit another issue to Elkowar, and considering I made myself look like
a complete nincompoop the last time I did that I'm not very excited.